refactor(authorization): move PermissionGrant + definitions to base module (ADR-020)
Per ADR-020, declarative primitives belong in the base module so non-EF
consumers (HTTP endpoints, jobs, tests) can use them without taking a
dependency on the EF Core runtime. Moves PermissionGrant entity,
PermissionGrantQueryDefinition and PermissionGrantExportDefinition out
of Granit.Authorization.EntityFrameworkCore into Granit.Authorization.
EntityFrameworkCore now references only QueryEngine/DataExchange
abstractions and registers the definitions via Add{Query,Export}Definition.
