+++ src/Granit.Authorization.EntityFrameworkCore/Extensions/AuthorizationEfCoreServiceCollectionExtensions.cs
@@ -1,11 +1,6 @@
 using Granit.Authorization;
 using Granit.Authorization.EntityFrameworkCore.DbContext;
-using Granit.Authorization.EntityFrameworkCore.Entities;
-using Granit.Authorization.EntityFrameworkCore.Exports;
-using Granit.Authorization.EntityFrameworkCore.Queries;
 using Granit.Authorization.EntityFrameworkCore.Stores;
-using Granit.DataExchange.Extensions;
-using Granit.QueryEngine.Extensions;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 
@@ -33,10 +28,6 @@ public static class AuthorizationEfCoreServiceCollectionExtensions
         // Replace the NullPermissionGrantStore registered by Granit.Authorization
         services.Replace(ServiceDescriptor.Scoped<IPermissionGrantStore,
             EfCorePermissionGrantStore<TContext>>());
-
-        // Query + Export definitions (ADR-020: owned by the base module).
-        services.AddQueryDefinition<PermissionGrant, PermissionGrantQueryDefinition>();
-        services.AddExportDefinition<PermissionGrant, PermissionGrantExportDefinition>();
 
         return services;
     }
